ingestion_service/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
confluent-kafka
requests
python-dotenv